List & list item handling changes
Keep track of current list type & index whenever a new [list] tag is
opened.

Modify tag regex so that [*] item tags can be picked up.  Added special
case in parser for '*', which increments the current list index and
attaches the current list type and the item's index to the token
information.  This allows filters to make use of the type & index, eg
for showing list bullets as numbers for a numbered list ([list=1]).